+++ src/Oscillator.cpp
@@ -80,6 +80,10 @@ struct Oscillator : VenomModule {
     TRI_ASIGN_PARAM,
     SQR_ASIGN_PARAM,
     SAW_ASIGN_PARAM,
+    
+    SINSHP_PARAM,
+    TRISHP_PARAM,
+    SAWSHP_PARAM,
 
     PARAMS_LEN
   };
@@ -165,13 +169,15 @@ struct Oscillator : VenomModule {
     MIX_RM_LIGHT,
     
     LIN_DC_LIGHT,
-
+    
     LIGHTS_LEN
   };
 
   bool disableOver[INPUTS_LEN]{};
   bool ringMod[5]{};
+  bool oldShpCV[4]{};
   float lvlScale[5]{0.1f, 0.1f, 0.1f, 0.1f, 0.1f};
+  float shpScale[4]{0.2f, 0.2f, 0.2f, 0.2f};
   bool softSync = false;
   bool alternate = false;
   bool lfo = false;
@@ -230,10 +236,15 @@ struct Oscillator : VenomModule {
 
     configSwitch<FixedSwitchQuantity>(MODE_PARAM, 0.f, 7.f, 0.f, "Frequency Mode", {"Audio frequency", "Low frequency", "0Hz carrier", "Triggered audio one shot", "Retriggered audio one shot", "Gated audio one shot", "Retriggered LFO one shot", "Gated LFO one shot"});
     configSwitch<FixedSwitchQuantity>(OVER_PARAM, 0.f, 5.f, 2.f, "Oversample", {"Off", "x2", "x4", "x8", "x16", "x32"});
-    configSwitch<FixedSwitchQuantity>(PW_PARAM, 0.f, 1.f, 0.f, "Pulse Width Range", {"Limited 3%-97%", "Full 0%-100%"});
-    configSwitch<FixedSwitchQuantity>(MIXSHP_PARAM, 0.f, 5.f, 0.f, "Mix Shape Mode", {"Sum (No shaping)", "Saturate Sum", "Fold Sum", "Average (No shaping)", "Saturate Average", "Fold Average"});
     configSwitch<FixedSwitchQuantity>(DC_PARAM,   0.f, 1.f, 0.f, "DC Block", {"Off", "On"});
     configButton(RESET_POLY_PARAM, "Reset polyphony count");
+
+    configSwitch<FixedSwitchQuantity>(SINSHP_PARAM, 0.f, 4.f, 0.f, "Sine Shape Mode", {"log/exp", "J-curve", "S-curve", "Rectify", "Normalized Rectify"});
+    configSwitch<FixedSwitchQuantity>(TRISHP_PARAM, 0.f, 4.f, 0.f, "Triangle Shape Mode", {"log/exp", "J-curve", "S-curve", "Rectify", "Normalized Rectify"});
+    configSwitch<FixedSwitchQuantity>(PW_PARAM, 0.f, 1.f, 0.f, "Square Shape Mode", {"Limited PWM 3%-97%", "Full PWM 0%-100%"});
+    configSwitch<FixedSwitchQuantity>(SAWSHP_PARAM, 0.f, 4.f, 0.f, "Saw Shape Mode", {"log/exp", "J-curve", "S-curve", "Rectify", "Normalized Rectify"});
+    configSwitch<FixedSwitchQuantity>(MIXSHP_PARAM, 0.f, 5.f, 0.f, "Mix Shape Mode", {"Sum (No shaping)", "Saturate Sum", "Fold Sum", "Average (No shaping)", "Saturate Average", "Fold Average"});
+
     
     configParam<FreqQuantity>(FREQ_PARAM, -4.f, 4.f, 0.f, "Frequency", " Hz");
     configParam(OCTAVE_PARAM, -4.f, 4.f, 0.f, "Octave");
@@ -319,6 +330,11 @@ struct Oscillator : VenomModule {
     lights[Oscillator::SIN_RM_LIGHT+indx].setBrightness(rm);
   }
   
+  void setOldShpCV(int indx, bool val) {
+    oldShpCV[indx] = val;
+    shpScale[indx] = val ? 0.1f : 0.2f;
+  }
+  
   void process(const ProcessArgs& args) override {
     VenomModule::process(args);
 
@@ -383,6 +399,13 @@ struct Oscillator : VenomModule {
       }
       softSync = !softSync;
     }
+    
+    int sinMode = static_cast<int>(params[SINSHP_PARAM].getValue());
+    int triMode = static_cast<int>(params[TRISHP_PARAM].getValue());
+    int sqrMode = static_cast<int>(params[PW_PARAM].getValue());
+    int sawMode = static_cast<int>(params[SAWSHP_PARAM].getValue());
+    int mixMode = static_cast<int>(params[MIXSHP_PARAM].getValue());
+    int mixType = mixMode % 3;
     
     bool procSin = outputs[SIN_OUTPUT].isConnected() || (outputs[MIX_OUTPUT].isConnected() && params[SIN_ASIGN_PARAM].getValue() != 1.f);
     bool procTri = outputs[TRI_OUTPUT].isConnected() || (outputs[MIX_OUTPUT].isConnected() && params[TRI_ASIGN_PARAM].getValue() != 1.f);
@@ -502,7 +525,8 @@ struct Oscillator : VenomModule {
               shapeIn[SIN] = shapeUpSample[s][SIN].process(shapeIn[SIN]);
             }
           } // preserve prior shapeIn[SIN] value
-          float_4 shape = clamp(shapeIn[SIN]*params[SIN_SHAPE_AMT_PARAM].getValue()*0.1f + params[SIN_SHAPE_PARAM].getValue(), -1.f, 1.f);
+          float_4 shape = clamp(shapeIn[SIN]*params[SIN_SHAPE_AMT_PARAM].getValue()*shpScale[SIN] + params[SIN_SHAPE_PARAM].getValue(), -1.f, 1.f);
+          float_4 shapeSign{};
           if (s==0 || inputs[SIN_PHASE_INPUT].isPolyphonic()) {
             phaseIn[SIN] = (o && !disableOver[SIN_PHASE_INPUT]) ? float_4::zero() : inputs[SIN_PHASE_INPUT].getPolyVoltageSimd<float_4>(c);
             if (procOver[SIN_PHASE_INPUT]){
@@ -514,7 +538,24 @@ struct Oscillator : VenomModule {
           sinPhasor = simd::fmod(sinPhasor, 1000.f);
           sinPhasor = simd::ifelse(sinPhasor<0.f, sinPhasor+1000.f, sinPhasor);
           sinPhasor = sinSimd_1000(sinPhasor);
-          sinOut[s] = crossfade(sinPhasor, ifelse(shape>0.f, 11.f*sinPhasor/(10.f*simd::abs(sinPhasor)+1.f), simd::sgn(sinPhasor)*simd::pow(sinPhasor,4)), ifelse(shape>0.f, shape, -shape))*5.f;
+          switch (sinMode) {
+            case 0:  // exp/log
+              sinOut[s] = crossfade(sinPhasor, ifelse(shape>0.f, 11.f*sinPhasor/(10.f*simd::abs(sinPhasor)+1.f), simd::sgn(sinPhasor)*simd::pow(sinPhasor,4)), ifelse(shape>0.f, shape, -shape))*5.f;
+              break;
+            case 1:  // J curve
+              sinOut[s] = (normSigmoid((sinPhasor+1.f)/2.f, -shape*0.9f)*2.f-1.f) * 5.f;
+              break;
+            case 2: // S curve
+              sinOut[s] = normSigmoid(sinPhasor, -shape*0.9f) * 5.f;
+              break;
+            default: // Rectify
+              shape = -shape;
+              float_4 shapeSign = simd::sgn(shape);
+              sinOut[s] = simd::ifelse(shapeSign==0, sinPhasor, -(shapeSign*simd::abs(-sinPhasor+shapeSign-shape)-shapeSign+shape));
+              if (sinMode==4 /*Normalized Rectify*/)
+                sinOut[s] = -((1+simd::abs(shape))*-sinOut[s]-shape);
+              sinOut[s] *= 5.f;
+          }
 
           if (s==0 || inputs[SIN_LEVEL_INPUT].isPolyphonic()) {
             levelIn[SIN] = (o && !disableOver[SIN_LEVEL_INPUT]) ? float_4::zero() : inputs[SIN_LEVEL_INPUT].getPolyVoltageSimd<float_4>(c);
@@ -522,7 +563,7 @@ struct Oscillator : VenomModule {
               if (o==0) levelIn[SIN] *= oversample;
               levelIn[SIN] = levelUpSample[s][SIN].process(levelIn[SIN]);
             }
-          } else levelIn[SIN] = levelIn[0][SIN];
+          } // else preserve prior levelIn[SIN] value
           level = levelIn[SIN]*params[SIN_LEVEL_AMT_PARAM].getValue()*lvlScale[SIN] + params[SIN_LEVEL_PARAM].getValue();
           if (params[SIN_ASIGN_PARAM].getValue()!=1) {
             mixOut[s] += sinOut[s] * level;
@@ -551,7 +592,7 @@ struct Oscillator : VenomModule {
               shapeIn[TRI] = shapeUpSample[s][TRI].process(shapeIn[TRI]);
             }
           } // else preserve prior shapeIn[TRI] value
-          float_4 shape = clamp(shapeIn[TRI]*params[TRI_SHAPE_AMT_PARAM].getValue()*0.1f + params[TRI_SHAPE_PARAM].getValue(), -1.f, 1.f);
+          float_4 shape = clamp(shapeIn[TRI]*params[TRI_SHAPE_AMT_PARAM].getValue()*shpScale[TRI] + params[TRI_SHAPE_PARAM].getValue(), -1.f, 1.f);
           if (s==0 || inputs[TRI_PHASE_INPUT].isPolyphonic()) {
             phaseIn[TRI] = (o && !disableOver[TRI_PHASE_INPUT]) ? float_4::zero() : inputs[TRI_PHASE_INPUT].getPolyVoltageSimd<float_4>(c);
             if (procOver[TRI_PHASE_INPUT]){
@@ -562,9 +603,27 @@ struct Oscillator : VenomModule {
           triPhasor = globalPhasor + (phaseIn[TRI]*params[TRI_PHASE_AMT_PARAM].getValue() + params[TRI_PHASE_PARAM].getValue()*2.f)*250.f + 250.f;
           triPhasor = simd::fmod(triPhasor, 1000.f);
           triPhasor = simd::ifelse(triPhasor<0.f, triPhasor+1000.f, triPhasor);
-          shape = simd::ifelse(triPhasor<500.f, shape, -shape);
+          if (triMode<=2) shape = simd::ifelse(triPhasor<500.f, shape, -shape);
           triPhasor = simd::ifelse(triPhasor<500.f, triPhasor*.002f, (1000.f-triPhasor)*.002f);
-          triOut[s] = crossfade(triPhasor, ifelse(shape>0.f, 11.f*triPhasor/(10.f*simd::abs(triPhasor)+1.f), simd::sgn(triPhasor)*simd::pow(triPhasor,4)), ifelse(shape>0.f, shape, -shape))*10.f-5.f;
+          switch (triMode) {
+            case 0:  // exp/log
+              triOut[s] = crossfade(triPhasor, ifelse(shape>0.f, 11.f*triPhasor/(10.f*simd::abs(triPhasor)+1.f), simd::sgn(triPhasor)*simd::pow(triPhasor,4)), ifelse(shape>0.f, shape, -shape))*10.f-5.f;
+              break;
+            case 1:  // J curve
+              triOut[s] = normSigmoid(triPhasor, -shape*0.8) * 10.f - 5.f;
+              break;
+            case 2: // S curve
+              triOut[s] = normSigmoid(triPhasor*2.f-1.f, -shape*0.8) * 5.f;
+              break;
+            default: // Rectify
+              shape = -shape;
+              float_4 shapeSign = simd::sgn(shape);
+              triOut[s] = triPhasor*2.f-1.f;
+              triOut[s] = simd::ifelse(shapeSign==0, triOut[s], -(shapeSign*simd::abs(-triOut[s]+shapeSign-shape)-shapeSign+shape));
+              if (triMode==4 /*Normalized Rectify*/)
+                triOut[s] = -((1+simd::abs(shape))*-triOut[s]-shape);
+              triOut[s] *= 5.f;
+          }
 
           if (s==0 || inputs[TRI_LEVEL_INPUT].isPolyphonic()) {
             levelIn[TRI] = (o && !disableOver[TRI_LEVEL_INPUT]) ? float_4::zero() : inputs[TRI_LEVEL_INPUT].getPolyVoltageSimd<float_4>(c);
@@ -572,7 +631,7 @@ struct Oscillator : VenomModule {
               if (o==0) levelIn[TRI] *= oversample;
               levelIn[TRI] = levelUpSample[s][TRI].process(levelIn[TRI]);
             }
-          } else levelIn[TRI] = levelIn[0][TRI];
+          } // else preserve prior levelIn[TRI] value
           level = levelIn[TRI]*params[TRI_LEVEL_AMT_PARAM].getValue()*lvlScale[TRI] + params[TRI_LEVEL_PARAM].getValue();
           if (params[TRI_ASIGN_PARAM].getValue()!=1){
             mixOut[s] += triOut[s] * level;
@@ -601,8 +660,8 @@ struct Oscillator : VenomModule {
               shapeIn[SQR] = shapeUpSample[s][SQR].process(shapeIn[SQR]);
             }
           } // else preserve prior shapeIn[SQR] value
-          float_4 flip = (shapeIn[SQR]*params[SQR_SHAPE_AMT_PARAM].getValue()*0.2f + params[SQR_SHAPE_PARAM].getValue() + 1.f) * 500.f;
-          if (!params[PW_PARAM].getValue()) flip = clamp( flip, 30.f, 970.f );
+          float_4 flip = (shapeIn[SQR]*params[SQR_SHAPE_AMT_PARAM].getValue()*shpScale[SQR] + params[SQR_SHAPE_PARAM].getValue() + 1.f) * 500.f;
+          if (!sqrMode) flip = clamp( flip, 30.f, 970.f );
           if (s==0 || inputs[SQR_PHASE_INPUT].isPolyphonic()) {
             phaseIn[SQR] = (o && !disableOver[SQR_PHASE_INPUT]) ? float_4::zero() : inputs[SQR_PHASE_INPUT].getPolyVoltageSimd<float_4>(c);
             if (procOver[SQR_PHASE_INPUT]){
@@ -621,7 +680,7 @@ struct Oscillator : VenomModule {
               if (o==0) levelIn[SQR] *= oversample;
               levelIn[SQR] = levelUpSample[s][SQR].process(levelIn[SQR]);
             }
-          } else levelIn[SQR] = levelIn[0][SQR];
+          } // else preserve prior levelIn[SQR] value
           level = levelIn[SQR]*params[SQR_LEVEL_AMT_PARAM].getValue()*lvlScale[SQR] + params[SQR_LEVEL_PARAM].getValue();
           if (params[SQR_ASIGN_PARAM].getValue()!=1){
             mixOut[s] += sqrOut[s] * level;
@@ -650,7 +709,7 @@ struct Oscillator : VenomModule {
               shapeIn[SAW] = shapeUpSample[s][SAW].process(shapeIn[SAW]);
             }
           } // else preserve prior shapeIn[SAW] value
-          float_4 shape = clamp(shapeIn[SAW]*params[SAW_SHAPE_AMT_PARAM].getValue()*0.1f + params[SAW_SHAPE_PARAM].getValue(), -1.f, 1.f);
+          float_4 shape = clamp(shapeIn[SAW]*params[SAW_SHAPE_AMT_PARAM].getValue()*shpScale[SAW] + params[SAW_SHAPE_PARAM].getValue(), -1.f, 1.f);
           if (s==0 || inputs[SAW_PHASE_INPUT].isPolyphonic()) {
             phaseIn[SAW] = (o && !disableOver[SAW_PHASE_INPUT]) ? float_4::zero() : inputs[SAW_PHASE_INPUT].getPolyVoltageSimd<float_4>(c);
             if (procOver[SAW_PHASE_INPUT]){
@@ -662,7 +721,25 @@ struct Oscillator : VenomModule {
           sawPhasor = simd::fmod(sawPhasor, 1000.f);
           sawPhasor = simd::ifelse(sawPhasor<0.f, sawPhasor+1000.f, sawPhasor);
           sawPhasor *= 0.001f;
-          sawOut[s] = crossfade(sawPhasor, ifelse(shape>0.f, 11.f*sawPhasor/(10.f*simd::abs(sawPhasor)+1.f), simd::sgn(sawPhasor)*simd::pow(sawPhasor,4)), ifelse(shape>0.f, shape, -shape))*10.f-5.f;
+          switch (sawMode) {
+            case 0:  // exp/log
+              sawOut[s] = crossfade(sawPhasor, ifelse(shape>0.f, 11.f*sawPhasor/(10.f*simd::abs(sawPhasor)+1.f), simd::sgn(sawPhasor)*simd::pow(sawPhasor,4)), ifelse(shape>0.f, shape, -shape))*10.f-5.f;
+              break;
+            case 1:  // J Curve
+              sawOut[s] = normSigmoid(sawPhasor, -shape*0.90) * 10.f - 5.f;
+              break;
+            case 2: // S Curve
+              sawOut[s] = normSigmoid(sawPhasor*2.f-1.f, -shape*0.85) * 5.f;
+              break;
+            default: // Rectify
+              shape = -shape;
+              float_4 shapeSign = simd::sgn(shape);
+              sawOut[s] = sawPhasor*2.f-1.f;
+              sawOut[s] = simd::ifelse(shapeSign==0, sawOut[s], -(shapeSign*simd::abs(-sawOut[s]+shapeSign-shape)-shapeSign+shape));
+              if (sawMode==4 /*Normalized Rectify*/)
+                sawOut[s] = -((1+simd::abs(shape))*-sawOut[s]-shape);
+              sawOut[s] *= 5.f;
+          }
 
           if (s==0 || inputs[SAW_LEVEL_INPUT].isPolyphonic()) {
             levelIn[SAW] = (o && !disableOver[SAW_LEVEL_INPUT]) ? float_4::zero() : inputs[SAW_LEVEL_INPUT].getPolyVoltageSimd<float_4>(c);
@@ -670,7 +747,7 @@ struct Oscillator : VenomModule {
               if (o==0) levelIn[SAW] *= oversample;
               levelIn[SAW] = levelUpSample[s][SAW].process(levelIn[SAW]);
             }
-          } else levelIn[SAW] = levelIn[0][SAW];
+          } // else preserve prior levelIn[SAW] value
           level = levelIn[SAW]*params[SAW_LEVEL_AMT_PARAM].getValue()*lvlScale[SAW] + params[SAW_LEVEL_PARAM].getValue();
           if (params[SAW_ASIGN_PARAM].getValue()!=1){
             mixOut[s] += sawOut[s] * level;
@@ -692,12 +769,11 @@ struct Oscillator : VenomModule {
         // Mix
         if (procMix) {
           int folds=10;
-          if (params[MIXSHP_PARAM].getValue() > 2.5) {
+          if (mixMode > 2) {
             mixOut[s] = simd::ifelse(mixDiv>0.f, mixOut[s]/mixDiv, mixOut[s]);
             folds=3;
           }
-          int typ = static_cast<int>(params[MIXSHP_PARAM].getValue()) % 3;
-          if (typ) {
+          if (mixType) {
             if (s==0 || inputs[MIX_SHAPE_INPUT].isPolyphonic()) {
               shapeIn[MIX] = (o && !disableOver[MIX_SHAPE_INPUT]) ? float_4::zero() : inputs[MIX_SHAPE_INPUT].getPolyVoltageSimd<float_4>(c);
               if (procOver[MIX_SHAPE_INPUT]){
@@ -706,10 +782,10 @@ struct Oscillator : VenomModule {
               }
             } // else preserve prior shapeIn[MIX] value
             float_4 drive = clamp(shapeIn[MIX]*params[MIX_SHAPE_AMT_PARAM].getValue() + params[MIX_SHAPE_PARAM].getValue()+1.f, 0.f, 3.f)*2.f + 1.f;
-            if (typ==1){
+            if (mixType==1){
               mixOut[s] = softClip<float_4>(mixOut[s]*2.f*drive)/2.f;
             }
-            if (typ==2){
+            if (mixType==2){
               mixOut[s] *= drive;
               float_4 clmp;
               for (int i=0; i<folds; i++){
@@ -796,9 +872,15 @@ struct Oscillator : VenomModule {
       json_array_append_new(array, json_boolean(disableOver[i]));
     json_object_set_new(rootJ, "disableOver", array);
     array = json_array();
-    for (int i=0; i<5; i++)
+    for (int i=0; i<5; i++){
       json_array_append_new(array, json_boolean(ringMod[i]));
+    }
     json_object_set_new(rootJ, "ringMod", array);
+    array = json_array();
+    for (int i=0; i<4; i++){
+      json_array_append_new(array, json_boolean(oldShpCV[i]));
+    }
+    json_object_set_new(rootJ, "oldShpCV", array);
     json_object_set_new(rootJ, "linDCCouple", json_boolean(linDCCouple));
     json_object_set_new(rootJ, "overParam", json_integer(params[OVER_PARAM].getValue()));
     return rootJ;
@@ -818,6 +900,16 @@ struct Oscillator : VenomModule {
       json_array_foreach(array, index, val){
         setRingMod( index, json_boolean_value(val));
       }
+    }
+    if ((array = json_object_get(rootJ, "oldShpCV"))) {
+      json_array_foreach(array, index, val){
+        setOldShpCV( index, json_boolean_value(val));
+      }
+    }
+    else {
+      setOldShpCV(SIN, true);
+      setOldShpCV(TRI, true);
+      setOldShpCV(SAW, true);
     }
     if ((val = json_object_get(rootJ, "linDCCouple"))) {
       linDCCouple = json_boolean_value(val);
@@ -858,8 +950,8 @@ struct OscillatorWidget : VenomWidget {
 
   struct PWSwitch : GlowingSvgSwitchLockable {
     PWSwitch() {
-      addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallOffButtonSwitch.svg")));
       addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallYellowButtonSwitch.svg")));
+      addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallBlueButtonSwitch.svg")));
     }
   };
 
@@ -871,6 +963,16 @@ struct OscillatorWidget : VenomWidget {
       addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallLightBlueButtonSwitch.svg")));
       addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallBlueButtonSwitch.svg")));
       addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallGreenButtonSwitch.svg")));
+    }
+  };
+  
+  struct ShpSwitch : GlowingSvgSwitchLockable {
+    ShpSwitch() {
+      addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallYellowButtonSwitch.svg")));
+      addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallOrangeButtonSwitch.svg")));
+      addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallPurpleButtonSwitch.svg")));
+      addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallLightBlueButtonSwitch.svg")));
+      addFrame(Svg::load(asset::plugin(pluginInstance,"res/smallBlueButtonSwitch.svg")));
     }
   };
 
@@ -927,6 +1029,23 @@ struct OscillatorWidget : VenomWidget {
     }
   };
   
+  struct ShapePort : PolyPort {
+    int portId;
+    void appendContextMenu(Menu* menu) override {
+      Oscillator* module = static_cast<Oscillator*>(this->module);
+      menu->addChild(new MenuSeparator);
+      menu->addChild(createBoolPtrMenuItem("Disable oversampling", "", &module->disableOver[portId]));
+      menu->addChild(createBoolMenuItem(
+        "20 VPP full range (old behavior)", "",
+        [=]() {return module->oldShpCV[portId-Oscillator::SIN_SHAPE_INPUT];},
+        [=](bool val) {
+          module->setOldShpCV(portId-Oscillator::SIN_SHAPE_INPUT, val);
+        }
+      ));
+      PolyPort::appendContextMenu(menu);
+    }
+  };
+  
   template <class TOverPort>
   TOverPort* createOverInputCentered(math::Vec pos, engine::Module* module, int inputId) {
     TOverPort* o = createInputCentered<TOverPort>(pos, module, inputId);
@@ -939,11 +1058,15 @@ struct OscillatorWidget : VenomWidget {
     setVenomPanel("Oscillator");
     
     addParam(createLockableParamCentered<ModeSwitch>(Vec(14.5f,37.5f), module, Oscillator::MODE_PARAM));
-    addParam(createLockableParamCentered<OverSwitch>(Vec(32.5f,37.5f), module, Oscillator::OVER_PARAM));
-    addParam(createLockableParamCentered<PWSwitch>(Vec(50.5f,37.5f), module, Oscillator::PW_PARAM));
-    addParam(createLockableParamCentered<MixShpSwitch>(Vec(68.5f,37.5f), module, Oscillator::MIXSHP_PARAM));
-    addParam(createLockableParamCentered<DCBlockSwitch>(Vec(86.5f,37.5f), module, Oscillator::DC_PARAM));
-    addParam(createLockableParamCentered<GlowingTinyButtonLockable>(Vec(315.5f,16.5f), module, Oscillator::RESET_POLY_PARAM));
+    addParam(createLockableParamCentered<OverSwitch>(Vec(35.5f,37.5f), module, Oscillator::OVER_PARAM));
+    addParam(createLockableParamCentered<DCBlockSwitch>(Vec(56.5f,37.5f), module, Oscillator::DC_PARAM));
+    addParam(createLockableParamCentered<GlowingTinyButtonLockable>(Vec(77.5f,37.5f), module, Oscillator::RESET_POLY_PARAM));
+
+    addParam(createLockableParamCentered<ShpSwitch>(Vec(130.0f,49.0f), module, Oscillator::SINSHP_PARAM));
+    addParam(createLockableParamCentered<ShpSwitch>(Vec(175.0f,49.0f), module, Oscillator::TRISHP_PARAM));
+    addParam(createLockableParamCentered<PWSwitch>(Vec(220.0f,49.0f), module, Oscillator::PW_PARAM));
+    addParam(createLockableParamCentered<ShpSwitch>(Vec(265.0f,49.0f), module, Oscillator::SAWSHP_PARAM));
+    addParam(createLockableParamCentered<MixShpSwitch>(Vec(310.0f,49.0f), module, Oscillator::MIXSHP_PARAM));
     
     addParam(createLockableParamCentered<RoundHugeBlackKnobLockable>(Vec(46.5f,93.5f), module, Oscillator::FREQ_PARAM));
     addParam(createLockableParamCentered<RotarySwitch<RoundBlackKnobLockable>>(Vec(29.f,157.f), module, Oscillator::OCTAVE_PARAM));
@@ -965,18 +1088,27 @@ struct OscillatorWidget : VenomWidget {
     addChild(createLightCentered<SmallLight<YellowRedLight<>>>(Vec(77.5f, 324.f), module, Oscillator::SYNC_LIGHT));
     
     float dx = 45.f;
-    float dy = 61.f;
+    float dy = 59.f;
     for (int y=0; y<4; y++) {
       for (int x=0; x<5; x++) {
-        addParam(createLockableParamCentered<RoundTinyBlackKnobLockable>(Vec(119.5f+dx*x,59.5f+dy*y), module, Oscillator::GRID_PARAM+y*10+x));
-        addParam(createLockableParamCentered<RoundTinyBlackKnobLockable>(Vec(140.5f+dx*x,59.5f+dy*y), module, Oscillator::GRID_PARAM+y*10+x+5));
-        if (y<3)
-          addInput(createOverInputCentered<OverPort>(Vec(130.f+dx*x,85.5f+dy*y), module, Oscillator::GRID_INPUT+y*5+x));
-        else
-          addInput(createOverInputCentered<LevelPort>(Vec(130.f+dx*x,85.5f+dy*y), module, Oscillator::GRID_INPUT+y*5+x));
-        addChild(createLightCentered<SmallLight<YellowRedLight<>>>(Vec(143.5f+dx*x, 74.f+dy*y), module, Oscillator::GRID_LIGHT+y*10+x*2));
+        addParam(createLockableParamCentered<RoundTinyBlackKnobLockable>(Vec(119.5f+dx*x,68.5f+dy*y), module, Oscillator::GRID_PARAM+y*10+x));
+        addParam(createLockableParamCentered<RoundTinyBlackKnobLockable>(Vec(140.5f+dx*x,68.5f+dy*y), module, Oscillator::GRID_PARAM+y*10+x+5));
+        switch (y) {
+          case 0:
+            if (x==2 || x==4)
+              addInput(createOverInputCentered<OverPort>(Vec(130.f+dx*x,94.5f+dy*y), module, Oscillator::GRID_INPUT+y*5+x));
+            else
+              addInput(createOverInputCentered<ShapePort>(Vec(130.f+dx*x,94.5f+dy*y), module, Oscillator::GRID_INPUT+y*5+x));
+            break;
+          case 3:
+            addInput(createOverInputCentered<LevelPort>(Vec(130.f+dx*x,94.5f+dy*y), module, Oscillator::GRID_INPUT+y*5+x));
+            break;
+          default:
+            addInput(createOverInputCentered<OverPort>(Vec(130.f+dx*x,94.5f+dy*y), module, Oscillator::GRID_INPUT+y*5+x));
+        }
+        addChild(createLightCentered<SmallLight<YellowRedLight<>>>(Vec(143.5f+dx*x, 83.f+dy*y), module, Oscillator::GRID_LIGHT+y*10+x*2));
         if (y==3)
-          addChild(createLightCentered<SmallSimpleLight<YellowLight>>(Vec(116.5f+dx*x, 74.f+dy*y), module, Oscillator::SIN_RM_LIGHT+x));
+          addChild(createLightCentered<SmallSimpleLight<YellowLight>>(Vec(116.5f+dx*x, 83.f+dy*y), module, Oscillator::SIN_RM_LIGHT+x));
       }
     }
     for (int x=0; x<4; x++) {