+++ src/VCAMix4Stereo.cpp
@@ -74,8 +74,9 @@ struct VCAMix4Stereo : MixBaseModule {
       "Bipolar linear band limited - CV unclamped", "Bipolar exponential band limited - CV unclamped"
     });
     configSwitch<FixedSwitchQuantity>(DCBLOCK_PARAM, 0.f, 3.f, 0.f, "Mix DC Block", {"Off", "Before clipping", "Before and after clipping", "After clipping"});
-    configSwitch<FixedSwitchQuantity>(CLIP_PARAM, 0.f, 6.f, 0.f, "Mix Clipping", {"Off", "Hard post-level", "Soft post-level", "Soft oversampled post-levl", 
-                                                                                         "Hard pre-level", "Soft pre-level", "Soft oversampled pre-level"});
+    configSwitch<FixedSwitchQuantity>(CLIP_PARAM, 0.f, 7.f, 0.f, "Mix Clipping", {"Off", "Hard post-level at 10V", "Soft post-level at 10V", "Soft oversampled post-levl at 10V", 
+                                                                                         "Hard pre-level at 10V", "Soft pre-level at 10V", "Soft oversampled pre-level at 10V",
+                                                                                         "Saturate (Soft oversampled post-level at 6V)"});
     configSwitch<FixedSwitchQuantity>(EXCLUDE_PARAM, 0.f, 1.f, 0.f, "Exclude Patched Outs from Mix", {"Off", "On"});
     configInput(LEFT_CHAIN_INPUT, "Left chain");
     configInput(RIGHT_CHAIN_INPUT, "Right chain")->description = "Normalled to left chain input";
@@ -461,7 +462,7 @@ struct VCAMix4Stereo : MixBaseModule {
       if (vcaMode >= 4)
         cv = leftVcaBandlimit[0][4][c/4].process(cv);
 
-      if (clip <= 3) {
+      if (clip <= 3 || clip == 7) {
         leftOut *= (params[MIX_LEVEL_PARAM].getValue()+offset)*scale*cv + postMixOff;
         rightOut *= (params[MIX_LEVEL_PARAM].getValue()+offset)*scale*cv + postMixOff;
       }
@@ -477,13 +478,13 @@ struct VCAMix4Stereo : MixBaseModule {
         leftOut = softClip(leftOut);
         rightOut = softClip(rightOut);
       }
-      if (clip == 3 || clip == 6){
+      if (clip == 3 || clip >= 6){
         for (int i=0; i<oversample; i++){
           leftOut = leftUpSample[c/4].process(i ? simd::float_4::zero() : leftOut*oversample);
-          leftOut = softClip(leftOut);
+          leftOut = (clip != 7) ? softClip(leftOut) : (softClip(leftOut*1.6667f) / 1.6667f);
           leftOut = leftDownSample[c/4].process(leftOut);
           rightOut = rightUpSample[c/4].process(i ? simd::float_4::zero() : rightOut*oversample);
-          rightOut = softClip(rightOut);
+          rightOut = (clip != 7) ? softClip(rightOut) : (softClip(rightOut*1.6667f) / 1.6667f);
           rightOut = rightDownSample[c/4].process(rightOut);
         }
       }
@@ -491,7 +492,7 @@ struct VCAMix4Stereo : MixBaseModule {
         leftOut = leftDcBlockAfterFilter[c/4].process(leftOut);
         rightOut = rightDcBlockAfterFilter[c/4].process(rightOut);
       }
-      if (clip > 3) {
+      if (clip > 3 && clip < 7) {
         leftOut *= (params[MIX_LEVEL_PARAM].getValue()+offset)*scale*cv + postMixOff;
         rightOut *= (params[MIX_LEVEL_PARAM].getValue()+offset)*scale*cv + postMixOff;
       }