+++ changelog.md
@@ -1,5 +1,30 @@
 # Venom Modules Changelog
 
+## 2.9.0 (2024-08-22)
+### New Modules
+- Benjolin Gates Expander
+- Benjolin Volts Expander
+- Thru
+
+### Enhancements
+- All Venom modules with expanders
+  - Added context menu options to add expanders
+- Mixers
+  - Added an extra clipping mode for saturation limiting at +/- 6V
+- Benjolin Oscillator
+  - Detune VCO 1 a tiny bit so PWM is not flat at initiation
+  - Add context menu option for unipolar clock input
+- VCO Lab
+  - Added shape mode options to Sin, Tri, and Saw waveforms
+  - Rearranged faceplate to be more logical
+  - Changed Sin, Tri, and Saw shape CV so 10 VPP covers full range, with option for old behavior
+  
+### Bug Fix
+- VCO Lab
+  - Fix Polyphonic CV level when level input is monophonic
+- Logic
+  - Fix locking for reuse and operator parameters
+
 ## 2.8.0 (2024-06-26)
 ### New Modules
 - Bay Input